Adds detection of InfoBlock data
The tool can now extract InfoBlock data from the decrypted bin and display them. It also display the signature of the firmware. In the future this signature will be used to validate the firmware with the OSIK key. InfoBlock data contains all sort of data like Version and specific info about the firmware installed. In the future the tool will be changed to accept decrypted bin to read the infoblock and validate the firmware.

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>